Ahora solo se manerajará 1 sola velocidad constante para la corrección de trayectoria.
Corrección de bugs al modificar la velocidad de los agentes físicos.

+++ src/Componentes/NXT/conexion/Bluethoot_conector.java
@@ -196,7 +196,7 @@ public abstract class Bluethoot_conector
         bt_env.enviar( Gestion_MensajesNXT.Enviar_MovimientoSimple_PARAR() );
     }
 
-    public void enviarVelocidad(int v)
+    public void enviarVelocidad(float v)
     {
         bt_env.enviar( Gestion_MensajesNXT.Enviar_SetVelocidad(v) );
     }
